modularizing (login check temporarily commented out)
