+++ src/app.js
@@ -39,7 +39,7 @@ var links = [
 
 // *************** ROUTING ***************
 
-app.config(['$routeProvider', 'checkLoggedin', function routeProvider($routeProvider, checkLoggedin) {
+app.config(['$routeProvider', function routeProvider($routeProvider) {
   $routeProvider
   .when('/login', {
     templateUrl: './views/login.html',
@@ -53,8 +53,8 @@ app.config(['$routeProvider', 'checkLoggedin', function routeProvider($routeProv
     templateUrl: './views/list.html',
     controller: 'RenderController',
     resolve: {
-      logincheck: function(checkLoggedin) {
-        return checkLoggedin.check();
+      logincheck: function() {
+        // return checkLoggedin.check();
       },
     }
   })
@@ -87,69 +87,6 @@ app.service('checkLoggedin',
     };
   }
 ]);
-
-app.factory('userSession', ['$location', '$http', '$rootScope', function ($location, $http, $rootScope) {
-  const login = function (loginData) {
-    return $http.post('/api/login', JSON.stringify(loginData))
-      .then(function (response) {
-        console.log('Login response: ', response);
-        if (loginData.email === response.data) {
-          console.log($location.path);
-          $rootScope.currentUser = response;
-          $location.path('/home');
-          console.log($location.path);
-        }
-      })
-      .catch(function (err) {
-        console.log('Login error: ', err);
-      });
-  };
-
-  const register = function (userRegData) {
-    return $http.post('/api/register', JSON.stringify(userRegData))
-    .then(function (response) {
-      console.log("Reg. response: ", response);
-      if (response.data.message) {
-        $location.url('/home');
-      }
-    })
-    .catch(function (err) {
-      console.log('Registration error: ', err);
-    });
-  };
-
-  return {
-    login: login,
-    register: register
-  }
-}]);
-
-// *********** CONTROLLERS ************
-app.controller('LoginController', ['$scope', 'sessionFactory', '$rootScope', function ($scope, sessionFactory, $rootScope) {
-  $scope.userLogin = function userLogin() {
-    var userLog = {
-      email: $scope.user.email,
-      password: $scope.user.password
-    };
-    sessionFactory.login(userLog);
-  };
-}]);
-
-app.controller('RegistrationController', ['$scope', 'sessionFactory', '$rootScope', function ($scope, sessionFactory, $rootScope) {
-  $scope.userRegister = function userRegister() {
-    if ($scope.user.password !== $scope.user.passwordRepeat) {
-      console.log("Error! Passwords don't match!")
-    }
-    else {
-      var userRegData = {
-        email: $scope.user.email,
-        password: $scope.user.password
-      };
-      console.log(userRegData);
-      sessionFactory.register(userRegData);
-    }
-  };
-}]);
 
 app.controller('RenderController', ['$scope', '$rootScope', '$http', '$location', function ($scope, $rootScope, $http, $location) {
   $scope.dummyLinks = links;
