+++ src/main.js
@@ -1,3 +1,6 @@
 require('angular');
 require('./scss/style.scss');
 require('./app.js');
+require('./usersession.factory.js');
+require('./logincontroller.controller.js');
+require('./registrationcontroller.controller.js');
